refactor: move functions used in register and edit to utils
